use companion object replace utils class
